fix: editor modularization
