+++ src/options/editor/alerts.ts
@@ -1,0 +1,12 @@
+export const showAlert = (message: string): void => {
+  const status = document.getElementById("status-css")!;
+  status.innerText = message;
+  status.classList.add("active");
+
+  setTimeout(() => {
+    status.classList.remove("active");
+    setTimeout(() => {
+      status.innerText = "";
+    }, 200);
+  }, 2000);
+};
