+++ src/options/editor/navigation.ts
@@ -1,0 +1,22 @@
+export const openEditCSS = (): void => {
+  const editCSS = document.getElementById("css");
+  const options = document.getElementById("options");
+  const themeContent = document.getElementById("themes-content");
+  if (editCSS && themeContent && options) {
+    editCSS.style.display = "block";
+    options.style.display = "none";
+    themeContent.style.display = "none";
+  }
+};
+
+export const openOptions = (): void => {
+  const editCSS = document.getElementById("css");
+  const options = document.getElementById("options");
+  const themeContent = document.getElementById("themes-content");
+
+  if (editCSS && themeContent && options) {
+    editCSS.style.display = "";
+    options.style.display = "";
+    themeContent.style.display = "";
+  }
+};
