+++ src/options/editor/types.ts
@@ -1,0 +1,35 @@
+export interface ModalOptions {
+  title: string;
+  message: string;
+  inputPlaceholder?: string;
+  inputValue?: string;
+  confirmText?: string;
+  cancelText?: string;
+  confirmDanger?: boolean;
+  showInput?: boolean;
+}
+
+export interface SaveResult {
+  success: boolean;
+  strategy?: "local" | "sync";
+  wasRetry?: boolean;
+  error?: any;
+}
+
+export interface BracketStackItem {
+  type: string;
+  from: number;
+}
+
+export interface ThemeCardOptions {
+  name: string;
+  author: string;
+  isCustom: boolean;
+  index: number;
+}
+
+declare global {
+  interface Window {
+    stylelint: any;
+  }
+}
